+++ paddle/fluid/operators/elementwise_add_op.h
@@ -74,7 +74,7 @@ class ElementwiseAddGradKernel : public framework::OpKernel<T> {
         }
         total += v1;
       }
-      printf("elementwise_add d_x: %f\n", total);
+      printf("elementwise_add d_x: %f\n", static_cast<double> total);
       //      std::cout << "elementwise_add dx: " << total << std::endl;
     }
 
@@ -90,7 +90,7 @@ class ElementwiseAddGradKernel : public framework::OpKernel<T> {
         }
         total += v1;
       }
-      printf("elementwise_add d_y: %f\n", total);
+      printf("elementwise_add d_y: %f\n", static_cast<double> total);
       //      std::cout << "elementwise_add dy: " << total << std::endl;
     }
   }
