+++ paddle/fluid/operators/elementwise_mul_op.h
@@ -78,7 +78,7 @@ class ElementwiseMulGradKernel : public framework::OpKernel<T> {
         }
         total += v1;
       }
-      printf("elementwise_mul d_x: %f\n", total);
+      printf("elementwise_mul d_x: %f\n", static_cast<double> total);
       //      std::cout << "elementwise_mul dx: " << total << std::endl;
     }
 
@@ -94,7 +94,7 @@ class ElementwiseMulGradKernel : public framework::OpKernel<T> {
         }
         total += v1;
       }
-      printf("elementwise_mul d_y: %f\n", total);
+      printf("elementwise_mul d_y: %f\n", static_cast<double> total);
       //      std::cout << "elementwise_mul dy: " << total << std::endl;
     }
   }
